Moved check_modules_installed to utils
check_modules_installed is useful in multiple modules, so utils is a
better place for is instead of ctapipe/io/unofficial/eventfilereader.py
