+++ ctapipe/utils/__init__.py
@@ -5,3 +5,4 @@ from .dynamic_class import dynamic_class_from_module
 from .table_interpolator import TableInterpolator
 from .datasets import (find_all_matching_datasets, get_table_dataset, get_dataset,
                        find_in_path)
+from .check_modules_installed import check_modules_installed
