+++ ctapipe/utils/tests/test_check_modules_installed.py
@@ -1,0 +1,7 @@
+from ctapipe.utils import check_modules_installed
+
+
+def test_check_modules_installed():
+    assert(not check_modules_installed(["unlikely_module_name"]))
+    assert(not check_modules_installed(["unlikely_module_name", 'numpy']))
+    assert(check_modules_installed(["numpy"]))
